Refactor the tests
