+++ glusterfs/tests/test_e2e.py
@@ -5,12 +5,13 @@ import pytest
 
 from datadog_checks.glusterfs import GlusterfsCheck
 
-from .common import CONFIG, EXPECTED_METRICS
+from .common import EXPECTED_METRICS
+
+pytestmark = pytest.mark.e2e
 
 
-@pytest.mark.e2e
-def test_e2e(dd_agent_check):
-    aggregator = dd_agent_check(CONFIG)
+def test_e2e(dd_agent_check, config):
+    aggregator = dd_agent_check(config)
 
     for metric in EXPECTED_METRICS:
         aggregator.assert_metric(metric)
